requirement.txt
torch==2.0.1
torchvision==0.15.2
numpy>=1.26.4
pillow>=9.0.0
matplotlib>=3.5.0
PyQt6>=6.4.0